feat: Implement focus(x, y)
